1501 User can not add an avatar

+++ Project/Zazo/Library/Classes/CRUD/Models/GetAvatarResponse.swift
@@ -18,7 +18,7 @@ struct GetAvatarResponse {
             case Avatar = "avatar"
         }
         
-        let timestamp: Int?
+        let timestamp: Int64?
         let useAsThumbnail: UseAsThumbnail?
     }
